+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - fixed: FIO OBT data not showing in received transaction memos
 - fixed: Fix broken Bity message signing
 - fixed: Return underLimit error for Simplex for very small quotes
+- fixed: FIO stake/unstake transactions displaying as sent/received transactions
 - removed: Removed pull-to-search feature from wallet list and transaction list scenes
 
 ## 4.11.0